+++ aws/postgres/locals.tf
@@ -13,4 +13,5 @@ locals {
   db_subnet_group_name         = "${var.db_name}-subnet"
   default_backup_retention     = var.backup_retention_day
   db_final_snapshot_identifier = "${local.identifier}-${formatdate("HH-mmaa", timestamp())}"
+  vpc_security_group_ids       = var.vpc_security_group_ids != null ? var.vpc_security_group_ids : [aws_security_group.this[0].id]
 }