+++ aws/postgres/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.15]() (2024-12-19)
+### Features
+* Add `var.custom_parameter_group_name` and `var.vpc_security_group_ids` to overwrite parameter group name and use exising security group ids.
+
 ## [0.1.14]() (2024-12-18)
 ### Features
 * Add `var.db_identifier` and update `identifier` naming logic from modules `main_db_instance` and `replica_db_instance`.